refactor(#3037): relocate SessionActivityResolver to services layer (backflow 35→33)
Move the session-activity resolver (SessionActivityResolver /
EffectiveSessionState / resolve_effective_state_with) out of
server/routes/session_activity.rs into a new services::session_activity module.

It is pure domain logic that only depends on crate::db::session_status and
crate::services::* (tmux_diagnostics, provider, tui_*, tmux_common) — it has
zero axum / server-type dependency, so it belongs beside the rest of the
service-layer infra rather than in the server route layer. No cycle is
introduced because the module already pointed only downward at db + services.

server/routes/session_activity.rs becomes a compat facade re-exporting
SessionActivityResolver, so the crate::server::routes::session_activity::* path
(mounted via the route module) keeps working unchanged for the route-layer
callers (v1.rs, stats.rs). Service consumers (services::agents::{query,turn})
and the db consumer (db::dispatched_sessions) are switched to the canonical
crate::services::session_activity path. Behavior-preserving; mirrors the
established monitoring_store facade pattern (#3211).

Service/server backflow: 35 → 33 (query.rs 2→1, turn.rs 2→1; the remaining
ref in each is the separate dto::agents bucket-4 concern). Also removes a
db→server backflow as a bonus.

Gates: cargo fmt --check clean; cargo check --lib clean (0 new warnings);
cargo test --lib services::agents passing; audit --check exit 0 (baseline
lowered, monotonic guard); generate_inventory_docs + check_agent_maintenance_docs
freshness passed; ci-script-checks.sh exit 0.

Refs #3037
