+++ scripts/audit_maintainability/baselines/service_server_backflow.json
@@ -3,13 +3,13 @@
   "rule": "service_server_backflow",
   "description": "No-regression baseline for src/services references to crate::server or super::server server-layer modules.",
   "captured_at": "2026-06-08",
-  "total_count": 35,
+  "total_count": 33,
   "files": {
     "src/services/agents/query.rs": {
-      "count": 2
+      "count": 1
     },
     "src/services/agents/turn.rs": {
-      "count": 2
+      "count": 1
     },
     "src/services/auto_queue/control_routes.rs": {
       "count": 1
