+++ src/services/agents/query.rs
@@ -5,7 +5,7 @@ use crate::server::dto::agents::{
     agent_office_json, agent_skill_json, build_channel_deeplinks, dedup_dispatched_sessions,
     dispatched_session_json, timeline_event_json,
 };
-use crate::server::routes::session_activity::SessionActivityResolver;
+use crate::services::session_activity::SessionActivityResolver;
 
 #[derive(Debug, Clone)]
 pub struct AgentDiagSession {
